BasePanel moved to shared
