--- SharePointFramework/ProjectWebParts/src/components/ProjectInformation/BasePanel/BasePanel.module.scss
@@ -1,6 +1,0 @@
-.root {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-  padding: 20px 24px;
-}
