--- SharePointFramework/ProjectWebParts/src/components/ProjectInformation/BasePanel/index.tsx
@@ -1,38 +1,0 @@
-import { Panel, PanelType } from '@fluentui/react'
-import strings from 'ProjectWebPartsStrings'
-import React, { FC } from 'react'
-import { useProjectInformationContext } from '../context'
-import { CLOSE_PANEL } from '../reducer'
-import { IBasePanelProps } from './types'
-import { FluentProvider, webLightTheme } from '@fluentui/react-components'
-import styles from './BasePanel.module.scss'
-
-export const BasePanel: FC<IBasePanelProps> = (props) => {
-  const context = useProjectInformationContext()
-
-  return (
-    <Panel
-      {...props}
-      isOpen={context.state.activePanel === props.$type}
-      onRenderBody={() => {
-        if (!props.onRenderBody) return null
-        return (
-          <FluentProvider theme={webLightTheme} className={styles.root}>
-            {props.onRenderBody()}
-          </FluentProvider>
-        )
-      }}
-      onDismiss={() => context.dispatch(CLOSE_PANEL())}
-    >
-      {props.children}
-    </Panel>
-  )
-}
-
-BasePanel.defaultProps = {
-  type: PanelType.medium,
-  isLightDismiss: true,
-  closeButtonAriaLabel: strings.CloseText
-}
-
-export * from './ClosePanelButton'
